changed floats to double

+++ gps_converter/gps_converter.c
@@ -15,32 +15,32 @@
 
 typedef struct gps_pos
 {
-	float lat;
-	float lon;
+	double lat;
+	double lon;
 } gps_pos;
 
-gps_pos cartesian_to_gps(float x, float y);
+gps_pos cartesian_to_gps(double x, double y);
 /*convert cartesian coordinates to gps coordinates*/
 
-float calc_dist_degrees(gps_pos* start, gps_pos* end);
-/*Calculate a distance between two positions given as floats*/
+double calc_dist_degrees(gps_pos* start, gps_pos* end);
+/*Calculate a distance between two positions given as doubles*/
 /*Result is returned as degrees*/
 
-float calc_start_course(gps_pos* start, gps_pos* end);
-/*Calculate starting course given start and end positions as floats*/
+double calc_start_course(gps_pos* start, gps_pos* end);
+/*Calculate starting course given start and end positions as doubles*/
 /*Result is returned as degrees*/
 
-float calc_end_pos_lat(gps_pos* start, float start_course, float dist);
+double calc_end_pos_lat(gps_pos* start, double start_course, double dist);
 /*Calculate end position latitude*/
 
-float d_sin(float deg);
+double d_sin(double deg);
 /*convert argument deg to radians and compute sine func*/
 
-float d_cos(float deg);
+double d_cos(double deg);
 /*convert argument deg to radians and compute cosine func*/
 
-float rad_to_deg(float rad);
-float deg_to_rad(float deg);
+double rad_to_deg(double rad);
+double deg_to_rad(double deg);
 
 int main(int argc, char *argv[])
 {
@@ -68,7 +68,7 @@ int main(int argc, char *argv[])
 	printf("UNIT TESTS:\n");
 
 
-	float a = 30;
+	double a = 30;
 
 	printf("sine of 30: %f\n", d_sin(a));
 
@@ -81,7 +81,7 @@ int main(int argc, char *argv[])
 	end.lat = 59.12;
 	end.lon = 22.43;
 
-	float dist = calc_dist_degrees(&start, &end);
+	double dist = calc_dist_degrees(&start, &end);
 	printf("distance is: %f\n", dist);
 
 
@@ -107,10 +107,10 @@ int main(int argc, char *argv[])
 /*calc_dist_degrees*/
 /*******************/
 
-float calc_dist_degrees(gps_pos* start, gps_pos* end)
+double calc_dist_degrees(gps_pos* start, gps_pos* end)
 {
 	
-	float val, dist;
+	double val, dist;
 
 	val = d_sin(start->lat) * d_sin(end->lat) + d_cos(start->lat) * d_cos(end->lat) * d_cos(start->lon - end->lon);
 	dist = rad_to_deg (acos(val));
@@ -128,11 +128,11 @@ float calc_dist_degrees(gps_pos* start, gps_pos* end)
 /*calc_start_course*/
 /*******************/
 
-float calc_start_course(gps_pos* start, gps_pos* end){
-	float num = d_sin(end->lat) - d_sin(start->lat) * d_cos(calc_dist_degrees(start, end));
-	float denom = d_cos(start->lat) * d_sin(calc_dist_degrees(start, end));
-	float res = num / denom;
-	float course = rad_to_deg(acos(res));
+double calc_start_course(gps_pos* start, gps_pos* end){
+	double num = d_sin(end->lat) - d_sin(start->lat) * d_cos(calc_dist_degrees(start, end));
+	double denom = d_cos(start->lat) * d_sin(calc_dist_degrees(start, end));
+	double res = num / denom;
+	double course = rad_to_deg(acos(res));
 	if (start->lon > end->lon)
 	{
 		return 360 - course;
@@ -148,10 +148,10 @@ float calc_start_course(gps_pos* start, gps_pos* end){
 /*calc_end_pos_lat*/
 /******************/
 
-float calc_end_pos_lat(gps_pos* start, float start_course, float dist){
-	float dist_degrees = (dist / 1852.0) / 60.0;
-	float left_exp = d_cos(start_course) * d_cos(start->lat) * d_sin(dist_degrees);
-	float right_exp = d_sin(start->lat) * d_cos(dist_degrees) * d_cos(dist_degrees);
+double calc_end_pos_lat(gps_pos* start, double start_course, double dist){
+	double dist_degrees = (dist / 1852.0) / 60.0;
+	double left_exp = d_cos(start_course) * d_cos(start->lat) * d_sin(dist_degrees);
+	double right_exp = d_sin(start->lat) * d_cos(dist_degrees) * d_cos(dist_degrees);
 	return rad_to_deg(asin(left_exp + right_exp));
 }
 
@@ -160,14 +160,14 @@ float calc_end_pos_lat(gps_pos* start, float start_course, float dist){
 /*calc_end_pos*/
 /**************/
 
-gps_pos calc_end_pos(gps_pos* start, float start_course, float dist){
+gps_pos calc_end_pos(gps_pos* start, double start_course, double dist){
 
 	gps_pos *end = (gps_pos*)malloc(sizeof(gps_pos));
 
-	float dist_degrees = (dist / 1852.0) / 60.0;
-	float end_pos_lat = calc_end_pos_lat(start, start_course, dist);
-	float num = d_cos(dist_degrees) - d_sin(start->lat) * d_sin(end_pos_lat);
-	float denom = d_cos(start->lat) * d_cos(end_pos_lat);
+	double dist_degrees = (dist / 1852.0) / 60.0;
+	double end_pos_lat = calc_end_pos_lat(start, start_course, dist);
+	double num = d_cos(dist_degrees) - d_sin(start->lat) * d_sin(end_pos_lat);
+	double denom = d_cos(start->lat) * d_cos(end_pos_lat);
 	
 	end->lat = end_pos_lat;
 
@@ -187,11 +187,11 @@ gps_pos calc_end_pos(gps_pos* start, float start_course, float dist){
 /*cartesian_to_gps*/
 /******************/
 
-gps_pos cartesian_to_gps(float x, float y){
+gps_pos cartesian_to_gps(double x, double y){
 
-	float dist = sqrt(pow(x, 2) + pow(y, 2));
+	double dist = sqrt(pow(x, 2) + pow(y, 2));
 	
-	float start_course = 90.0 - rad_to_deg(asin(y/dist));
+	double start_course = 90.0 - rad_to_deg(asin(y/dist));
 	
 	gps_pos* start = (gps_pos*)malloc(sizeof(gps_pos));
 
@@ -205,8 +205,8 @@ gps_pos cartesian_to_gps(float x, float y){
 
 
 
-float d_sin(float deg) { return (sin(deg_to_rad (deg))); }
-float d_cos(float deg) { return (cos(deg_to_rad(deg))); }
+double d_sin(double deg) { return (sin(deg_to_rad (deg))); }
+double d_cos(double deg) { return (cos(deg_to_rad(deg))); }
 
-float rad_to_deg(float rad) { return rad * 180 / PI; }
-float deg_to_rad(float deg) { return deg * PI / 180; }
+double rad_to_deg(double rad) { return rad * 180 / PI; }
+double deg_to_rad(double deg) { return deg * PI / 180; }
